+++ bad-exports.txt
@@ -11,15 +11,10 @@ Sending 3 files to free worker...
 Sending 3 files to free worker...
 Sending 3 files to free worker...
 Sending 1 files to free worker...
-[error:bad-export] app/routes/bookmarks/index.tsx:113 BookmarkEntryComponent
-[error:bad-export] app/routes/bookmarks/index.tsx:185 MiniPlayer
-[error:bad-export] app/routes/bookmarks/index.tsx:430 BookmarksMenuItems
-[error:bad-export] app/routes/videos/$id.tsx:60 findCurrentEntry
-[error:bad-export] app/routes/videos/$id.tsx:510 CaptionEntryComponent
 All done. 
 Results: 
 0 errors
 0 unmodified
 31 skipped
 0 ok
-Time elapsed: 0.973seconds 
+Time elapsed: 0.950seconds
